Update on "Use std::isnan instead of self-comparison."
